Move all pipe stuff to same package
