+++ server-extension/src/main/java/hsy/pipe/GetPipesWithParams.java
@@ -1,6 +1,4 @@
-package hsy.pipe.actions;
-
-import hsy.pipe.helpers.Helpers;
+package hsy.pipe;
 
 import java.io.BufferedReader;
 import java.io.IOException;
