+++ server-extension/src/main/java/hsy/pipe/TagPipeActionHandler.java
@@ -1,9 +1,7 @@
-package hsy.pipe.actions;
+package hsy.pipe;
 
 import org.json.JSONObject;
 
-import hsy.pipe.domain.TagPipeConfiguration;
-import hsy.pipe.helpers.TagPipeHelper;
 import fi.nls.oskari.annotation.OskariActionRoute;
 import fi.nls.oskari.control.ActionException;
 import fi.nls.oskari.control.ActionParameters;
