+++ server-extension/src/main/java/hsy/pipe/TagPipeConfiguration.java
@@ -1,4 +1,4 @@
-package hsy.pipe.domain;
+package hsy.pipe;
 
 import org.json.JSONObject;
